Develop

+++ library/libraries/LI_Session.php
@@ -4,6 +4,8 @@ class LI_Session extends CI_Session {
 
 	private $_md5_encrypt = false;
 	private $_redis = null;
+	private $_redis_select = false;
+	private $_use_db = true;
 
 	public function __construct($params = array())
 	{
@@ -39,7 +41,7 @@ class LI_Session extends CI_Session {
 		// Are we using a database?  If so, load it
 		if ($this->sess_use_database === TRUE AND $this->sess_table_name != '')
 		{
-			$this->CI->load->database();
+			if($this->_use_db) $this->CI->load->database('',true);
 		}
 
 		// Set the "now" time.  Can either be GMT or server time, based on the
@@ -81,11 +83,13 @@ class LI_Session extends CI_Session {
 
 	private function _redis_load()
 	{
-		if($this->_redis !== false && !$this->CI->input->cookie('_nrd') && extension_loaded('redis'))
+		$_nrd = $this->CI->input->cookie('_nrd');
+		if($this->_redis !== false && !$_nrd && extension_loaded('redis'))
 		{
 			$this->CI->config->load('redis', TRUE, TRUE);
 			$redis_config = $this->CI->config->item('redis');
 			$config = $redis_config['redis_default'];
+			$this->_redis_select=$redis_config['redis_db']['session'];
 			//$config['timeout'] = 0.1;
 
 			$redis = new Redis();
@@ -103,7 +107,7 @@ class LI_Session extends CI_Session {
 				try
 				{
 					$redis->auth($config['password']);
-					$redis->select($redis_config['redis_db']['session']);
+					$redis->select($this->_redis_select);
 				}
 				catch (RedisException $e)
 				{
@@ -118,7 +122,7 @@ class LI_Session extends CI_Session {
 			}
 			$this->_redis = $redis;
 		}
-		else
+		else if(!$_nrd)
 		{
 			$this->CI->input->set_cookie('_nrd','1',0);
 		}
@@ -126,26 +130,53 @@ class LI_Session extends CI_Session {
 
 	private function _redis_set($key, $value, $ttl)
 	{
-		if($this->_redis) return $this->_redis->set($key, $value, $ttl);
-		else log_message('error_tizi', '100706:session redis set failed');
+		if($this->_redis && $this->_redis_select) 
+		{
+			$this->_redis->select($this->_redis_select);
+			return $this->_redis->set($key, $value, $ttl);
+		}
+		else 
+		{
+			log_message('error_tizi', '100706:session redis set failed');
+		}
 	}
 
 	private function _redis_get($key)
 	{
-		if($this->_redis) return $this->_redis->get($key);
-		else log_message('error_tizi', '100707:session redis get failed');
+		if($this->_redis && $this->_redis_select) 
+		{
+			$this->_redis->select($this->_redis_select);
+			return $this->_redis->get($key);
+		}
+		else 
+		{
+			log_message('error_tizi', '100707:session redis get failed');
+		}
 	}
 
 	private function _redis_del($key)
 	{
-		if($this->_redis) return $this->_redis->del($key);
-		else log_message('error_tizi', '100708:session redis del failed');
+		if($this->_redis && $this->_redis_select) 
+		{
+			$this->_redis->select($this->_redis_select);
+			return $this->_redis->del($key);
+		}
+		else 
+		{
+			log_message('error_tizi', '100708:session redis del failed');
+		}
 	}
 
 	private function _redis_close()
 	{
-		if($this->_redis) return $this->_redis->close();
-		else log_message('error_tizi', '100709:session redis close failed');
+		if($this->_redis) 
+		{
+			return $this->_redis->close();
+		}
+		else 
+		{
+			log_message('error_tizi', '100709:session redis close failed');
+		}
 	}
 
 	function sess_read()
@@ -202,22 +233,33 @@ class LI_Session extends CI_Session {
 				$userdata = $this->_redis_get($session_id);
 				if(!empty($userdata)) $session = json_decode($userdata,true);
 			}
-			
-			if(empty($session))
+			if(empty($session)&&$this->_use_db)
 			{
+				$this->CI->load->database('',true);
 				$this->CI->db->where('session_id',$session_id);
 				$query = $this->CI->db->get($this->sess_table_name);
 
 				// No result?  Kill it!
 				if ($query->num_rows() == 0)
 				{
-					log_message('trace_tizi','session_destroy_unload_session');
-					$this->sess_destroy();
-					return FALSE;
+					$session = array();
 				}
+				else
+				{
+					// Is there custom data?  If so, add it to the main session array
+					$session = $query->row_array();
+					if($this->_redis)
+					{
+						$this->_redis_set($session['session_id'],json_encode($session),$this->sess_expiration);
+					}
+				}
+			}
 
-				// Is there custom data?  If so, add it to the main session array
-				$session = $query->row_array();
+			if(empty($session))
+			{
+				log_message('trace_tizi','session_destroy_unload_session');
+				$this->sess_destroy();
+				return FALSE;
 			}
 
 			if (isset($session['user_data']) AND $session['user_data'] != '')
@@ -314,14 +356,20 @@ class LI_Session extends CI_Session {
 		}
 
 		// Run the update query
-		$this->CI->db->where('session_id', $this->userdata['session_id']);
-		$this->CI->db->update($this->sess_table_name, array('last_activity' => $this->userdata['last_activity'], 'user_data' => $custom_userdata));
-
-		$userdata = $this->_redis_get($this->userdata['session_id']);
-		$userdata = json_decode($userdata,true);
-		$userdata['last_activity'] = $this->userdata['last_activity'];
-		$userdata['user_data'] = $custom_userdata;
-		$this->_redis_set($this->userdata['session_id'],json_encode($userdata),$this->sess_expiration);
+		if($this->_redis)
+		{
+			$userdata = $this->_redis_get($this->userdata['session_id']);
+			$userdata = json_decode($userdata,true);
+			$userdata['last_activity'] = $this->userdata['last_activity'];
+			$userdata['user_data'] = $custom_userdata;
+			$this->_redis_set($this->userdata['session_id'],json_encode($userdata),$this->sess_expiration);
+		}
+		else if($this->_use_db)
+		{
+			$this->CI->load->database('',true);
+			$this->CI->db->where('session_id', $this->userdata['session_id']);
+			$this->CI->db->update($this->sess_table_name, array('last_activity' => $this->userdata['last_activity'], 'user_data' => $custom_userdata));
+		}
 		// Write the cookie.  Notice that we manually pass the cookie data array to the
 		// _set_cookie() function. Normally that function will store $this->userdata, but
 		// in this case that array contains custom data, which we do not want in the cookie.
@@ -353,8 +401,15 @@ class LI_Session extends CI_Session {
 		// Save the data to the DB if needed
 		if ($this->sess_use_database === TRUE)
 		{
-			$this->CI->db->query($this->CI->db->insert_string($this->sess_table_name, $this->userdata));
-			$this->_redis_set($this->userdata['session_id'],json_encode($this->userdata),$this->sess_expiration);
+			if($this->_redis)
+			{
+				$this->_redis_set($this->userdata['session_id'],json_encode($this->userdata),$this->sess_expiration);
+			}
+			else if($this->_use_db) 
+			{
+				$this->CI->load->database('',true);
+				$this->CI->db->query($this->CI->db->insert_string($this->sess_table_name, $this->userdata));
+			}
 		}
 
 		// Write the cookie
@@ -402,15 +457,21 @@ class LI_Session extends CI_Session {
 				$cookie_data[$val] = $this->userdata[$val];
 			}
 
-			$this->CI->db->query($this->CI->db->update_string($this->sess_table_name, array('last_activity' => $this->now, 'session_id' => $new_sessid), array('session_id' => $old_sessid)));
-			
-			$userdata = $this->_redis_get($old_sessid);
-			$userdata = json_decode($userdata,true);
-			$userdata['last_activity'] = $this->now;
-			$userdata['session_id'] = $new_sessid;
-			$this->_redis_set($new_sessid,json_encode($userdata),$this->sess_expiration);
-			//$this->_redis_del($old_sessid);
-			$this->_redis_set($old_sessid,json_encode($userdata),120);
+			if($this->_redis)
+			{
+				$userdata = $this->_redis_get($old_sessid);
+				$userdata = json_decode($userdata,true);
+				$userdata['last_activity'] = $this->now;
+				$userdata['session_id'] = $new_sessid;
+				$this->_redis_set($new_sessid,json_encode($userdata),$this->sess_expiration);
+				//$this->_redis_del($old_sessid);
+				$this->_redis_set($old_sessid,json_encode($userdata),120);
+			}
+			else if($this->_use_db) 
+			{
+				$this->CI->load->database('',true);
+				$this->CI->db->query($this->CI->db->update_string($this->sess_table_name, array('last_activity' => $this->now, 'session_id' => $new_sessid), array('session_id' => $old_sessid)));
+			}
 		}
 
 		// Write the cookie
@@ -422,21 +483,27 @@ class LI_Session extends CI_Session {
 		// Kill the session DB row
 		if ($this->sess_use_database === TRUE && isset($this->userdata['session_id']))
 		{
-			$this->CI->db->where('session_id', $this->userdata['session_id']);
-			$this->CI->db->delete($this->sess_table_name);
-
-			$this->_redis_del($this->userdata['session_id']);
+			if($this->_redis) 
+			{
+				$this->_redis_del($this->userdata['session_id']);
+			}
+			else if($this->_use_db)
+			{
+				$this->CI->load->database('',true);
+				$this->CI->db->where('session_id', $this->userdata['session_id']);
+				$this->CI->db->delete($this->sess_table_name);
+			}
 		}
 
 		// Kill the cookie
 		setcookie(
-					$this->sess_cookie_name,
-					addslashes(serialize(array())),
-					($this->now - 31500000),
-					$this->cookie_path,
-					$this->cookie_domain,
-					0
-				);
+			$this->sess_cookie_name,
+			addslashes(serialize(array())),
+			($this->now - 31500000),
+			$this->cookie_path,
+			$this->cookie_domain,
+			0
+		);
 
 		// Kill session data
 		$this->userdata = array();
@@ -474,13 +541,45 @@ class LI_Session extends CI_Session {
 
 		// Set the cookie
 		setcookie(
-					$this->sess_cookie_name,
-					$cookie_data,
-					$expire,
-					$this->cookie_path,
-					$this->cookie_domain,
-					$this->cookie_secure
-				);
+			$this->sess_cookie_name,
+			$cookie_data,
+			$expire,
+			$this->cookie_path,
+			$this->cookie_domain,
+			$this->cookie_secure
+		);
+	}
+
+	function _sess_gc()
+	{
+		if ($this->sess_use_database != TRUE)
+		{
+			return;
+		}
+
+		if ($this->_redis)
+		{
+			return;
+		}
+		if (!$this->_use_db)
+		{
+			return;
+		}
+
+		srand(time());
+		if ((rand() % 100) < $this->gc_probability)
+		{
+			$expire = $this->now - $this->sess_expiration;
+
+			if($this->_use_db)
+			{
+				$this->CI->load->database('',true);
+				$this->CI->db->where("last_activity < {$expire}");
+				$this->CI->db->delete($this->sess_table_name);
+			}
+
+			log_message('debug', 'Session garbage collection performed.');
+		}
 	}
 
 }
